:pencil: Change status character
Since Ubuntu version has been updated to 18, the status character
stopped being colored. So, the character has been changed to a
simple '*' in order to keep coloring as it was used to.

Maybe we could check if the system running the fancygit is Ubuntu or
not, and so we could change the status character according to the
system.