:pencil: Change the status that indicates changes
The old status was used to be the number of file changes followed by the
m letter. So, for example, if we had 5 files changed, it would be like
5m.

Now, it doesn't show the number of files changed, just the status
informing that there are changes and it is +-. As +- is already part of
git diff when you check out the changes for what have been added and
what have been removed, it seems to be the perfect indicator status for
that purpose.