Add a new font in order to keep compatibility
It fixes #70

Bacause of the new font, some changes have to made on icons.

- Update virtualenv status icon.
- Change stash status icon.

The github icon has changed also but it's more like the old one.

This changes have to be made in order to keep showing the same status
icons when using the Sauce Code font (the new one).