:sparkles: Add a new update manager
The update manager now checks for updates asynchronously.
Another change here is that it will check for the last release on
github and compare it against the last release on local machine.
It notifies for new updates when the last release on github is
greater than the one on local machine.