:sparkles: Separate themes from color schemes
So far FancyGit had a couple of commands to switching between themes
(prompt styles). This has been changed. Now we have themes and color
schemes commands.

The concepts have been separated in order to reach a better structure.

Themes are related to how things are placed on the PS1 prompt.
Color schemes are related to how the prompt will be colored.

We can switch themes by typing: fancygit --theme-<name>

But if all we want to do is switching the colors, so we can type:
fancygit --color-scheme-<name>